move errors to its corresponding place
