+++ dial9-tokio-telemetry/src/lib.rs
@@ -38,4 +38,4 @@ mod telemetry_runtime;
 
 pub use current_config::{Dial9Config, Dial9ConfigBuilder, Dial9ConfigBuilderError, MissingFields};
 pub use dial9_macro::main;
-pub use telemetry_runtime::TelemetryRuntime;
+pub use telemetry_runtime::{TelemetryRuntime, TelemetryRuntimeError};
